Restricted pages

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -14,6 +16,9 @@ from multiprocessing import context
 
 
 def loginPage(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
         username = request.POST.get('username')
@@ -62,7 +67,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     
@@ -76,9 +81,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user !=room.host:
+        return HttpResponse("You are not allowed here")
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -88,8 +97,15 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id = pk)
+    
+    if request.user !=room.host:
+        return HttpResponse("You are not allowed here")
+    
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
